sql-agent-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
